Moved to 'base' module logic of the game
